Finish data_clean first draft
data_clean is now essentially done for the time being; I've removed all
the features and interpolated all the values--in other words, this data
is ready for use in machine learning. There are a few possibilities for
improvement after this, such as taking the date into account, but we do
have a complete dataset so we can get something working at the very
least.